LayoutTests/webanimations/animation-effect-timing.html

<script src="../resources/js-test-pre.js"></script>
<script>

description("Check that the AnimationEffectTiming interface is defined and does not expose a constructor.");

debug("The AnimationEffectTiming interface is defined.");
shouldBeDefined("AnimationEffectTiming");

debug("");
debug("The AnimationEffectTiming interface cannot be constructed.");
shouldThrow("new AnimationEffectTiming");

debug("");

</script>
<script src="../resources/js-test-post.js"></script>
