LayoutTests/webanimations/animation-effect-timing-expected.txt

Check that the AnimationEffectTiming interface is defined and does not expose a constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The AnimationEffectTiming interface is defined.
PASS AnimationEffectTiming is defined.

The AnimationEffectTiming interface cannot be constructed.
PASS new AnimationEffectTiming threw exception TypeError: function is not a constructor (evaluating 'new AnimationEffectTiming').

PASS successfullyParsed is true

TEST COMPLETE
